refactor: remove legacy SSE app code
