+++ pyproject.toml
@@ -22,9 +22,7 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-profiling = [
-  "austin-dist>=3.7.0",
-]
+profiling = ["austin-dist>=3.7.0"]
 
 [dependency-groups]
 dev = [
@@ -103,11 +101,12 @@ venvPath = "."
 venv = ".venv"
 include = ["."]
 exclude = [
-    "libraries/griptape_nodes_library/workflows/templates/**/*",
-    ".venv",
-    "**/node_modules",
-    "**/__pycache__",
-    "**/.*"
+  "libraries/griptape_nodes_library/workflows/templates/**/*",
+  ".venv",
+  "GriptapeNodes",
+  "**/node_modules",
+  "**/__pycache__",
+  "**/.*",
 ]
 pythonVersion = "3.12"
 reportIncompatibleMethodOverride = false
